Bring cleanup code up to its usage point
